requirements.txt
cloudinary==1.26.0
Flask==2.0.2
Flask_Login==0.5.0
Flask_SQLAlchemy==2.5.1
Flask_WTF==0.15.1
gunicorn==20.1.0
psycopg2==2.9.1
python-dotenv==0.19.1
python_bcrypt==0.3.2
SQLAlchemy==1.4.26
Werkzeug==2.0.2
WTForms==2.3.3
